add link posts to listItems
