+++ components/post/listItem.tsx
@@ -1,13 +1,11 @@
-import { PostType } from "@prisma/client";
 import { Box, Heading } from "@chakra-ui/react";
 import { useRouter } from "next/router";
-import Image from "next/image";
 
 import { PostWithCommentsCountAndForum } from "lib/api";
-import { Markdown } from "../markdown";
-import { PostMeta } from "../post/meta";
-import { Panel } from "../panel";
-import { PostButtons } from "./buttons";
+import { Panel } from "components/panel";
+import { PostMeta } from "components/post/meta";
+import { PostButtons } from "components/post/buttons";
+import { PostContent } from "components/post/content";
 
 interface PostListItemProps {
   post: PostWithCommentsCountAndForum;
@@ -15,52 +13,6 @@ interface PostListItemProps {
 
 export const PostListItem = ({ post }: PostListItemProps) => {
   const router = useRouter();
-
-  function renderBody() {
-    switch (post.postType) {
-      case PostType.IMAGE: {
-        return (
-          <Box
-            position="relative"
-            width="100%"
-            maxHeight="512px"
-            sx={{
-              "&:before": {
-                content: '""',
-                display: "block",
-                paddingBottom: "100%",
-              },
-            }}
-          >
-            <Image
-              fill
-              src={post.uri}
-              alt="post image"
-              style={{
-                objectFit: "cover",
-                maxWidth: "100%",
-                maxHeight: "100%",
-              }}
-            />
-          </Box>
-        );
-      }
-
-      case PostType.TEXT:
-      default: {
-        return (
-          <Box position="relative" maxHeight="250px">
-            <Markdown>{post.body ?? ""}</Markdown>
-            <Box
-              position="absolute"
-              inset={0}
-              background="linear-gradient(to bottom, transparent 100px, var(--chakra-colors-onda-950))"
-            />
-          </Box>
-        );
-      }
-    }
-  }
 
   return (
     <Panel
@@ -80,7 +32,7 @@ export const PostListItem = ({ post }: PostListItemProps) => {
         <Heading my="4" fontSize="2xl" fontWeight="semibold">
           {post.title}
         </Heading>
-        {renderBody()}
+        <PostContent type={post.postType} body={post.body} uri={post.uri} />
       </Box>
       <PostButtons post={post} />
     </Panel>
